--- assets2/icons/bigDelete.ts
@@ -1,12 +1,0 @@
-import { Icon } from "../../libs/type/Icon";
-
-const bigDelete: Icon = {
-  d:
-    "M18.67,22.26,11.5,15.09,4.33,22.26A2.54,2.54,0,1,1,.74,18.67h0L7.91,11.5.74,4.33A2.54,2.54,0,0,1,.74.74h0a2.54,2.54,0,0,1,3.59,0L11.5,7.91,18.67.74a2.54,2.54,0,0,1,3.59,0h0a2.54,2.54,0,0,1,0,3.59L15.09,11.5l7.17,7.17a2.54,2.54,0,0,1,0,3.59,2.55,2.55,0,0,1-3.59,0Z",
-  width: 23,
-  height: 23,
-  fill: "#FF587E",
-  viewBox: "0 0 23 23",
-};
-
-export default bigDelete;
